+++ examples/fm.rs
@@ -12,10 +12,13 @@ fn main() {
     assert!(ff.fd() >= 0);
 
     let fm = fmmap::FrozenMMap::new(ff, fmmap::FMCfg::new(module_id)).expect("mmap");
+    let (_, epoch) = fm.with_write::<u64, _>(0, |v| *v = 0xDEADC0DE).unwrap();
 
-    fm.with_write::<u64, _>(0, |v| *v = 0xDEADC0DE).unwrap();
-    fm.sync().expect("sync");
-
-    let value = fm.with_read::<u64, u64>(0, |v| *v).unwrap();
-    assert_eq!(value, 0xDEADC0DE);
+    match fm.wait_for_durability(epoch) {
+        Ok(_) => {
+            let value = fm.with_read::<u64, u64>(0, |v| *v).unwrap();
+            assert_eq!(value, 0xDEADC0DE);
+        }
+        Err(e) => panic!("{e}"),
+    }
 }